refactor: split Logger and assert from AsstUtils
