+++ src/MeoAssistant/MeoAssistant.vcxproj
@@ -124,6 +124,8 @@
     <ClInclude Include="Utils\AsstRanges.hpp" />
     <ClInclude Include="Utils\AsstTypes.h" />
     <ClInclude Include="Utils\AsstUtils.hpp" />
+    <ClInclude Include="Utils\Demangle.hpp" />
+    <ClInclude Include="Utils\Meta.hpp" />
     <ClInclude Include="Utils\Logger.hpp" />
     <ClInclude Include="Utils\NoWarningCV.h" />
     <ClInclude Include="Utils\NoWarningCVMat.h" />
