+++ src/MeoAssistant/MeoAssistant.vcxproj.filters
@@ -653,6 +653,12 @@
     <ClInclude Include="Utils\AsstTypes.h">
       <Filter>源文件\Utils</Filter>
     </ClInclude>
+    <ClInclude Include="Utils\Meta.hpp">
+      <Filter>源文件\Utils</Filter>
+    </ClInclude>
+    <ClInclude Include="Utils\Demangle.hpp">
+      <Filter>源文件\Utils</Filter>
+    </ClInclude>
     <ClInclude Include="Utils\NoWarningCV.h">
       <Filter>源文件\Utils</Filter>
     </ClInclude>
